requirements.txt
requests
python_telegram_bot
redis
cachetools
dateutils
leather
cairosvg
dateparser
bidict